requirements.txt
fastmcp>=2.0.0
python-dotenv>=1.0.1
pypdf>=4.2.0
pdfplumber>=0.11.0
pymupdf>=1.24.0
chromadb>=0.5.5
sentence-transformers>=3.0.0
numpy>=1.26.0
pydantic>=2.7.0
rich>=13.7.0
typer>=0.12.0
boto3>=1.34.0
spacy>=3.7.0
httpx>=0.27.0

# Brain + MCP tools
langchain>=0.2.0
langchain-core>=0.2.0
langchain-openai>=0.2.0
openai>=1.40.0
langchain-mcp-adapters>=0.0.1

# NEW UI backend
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
python-multipart>=0.0.9
anyio>=4.4.0
